requirements.txt
aiogram==3.9.0
pyzbar==0.1.9
opencv-python==4.9.0.80
python-barcode==0.15.1
Pillow==10.2.0
ultralytics==8.1.34
torch